requirements.txt
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.66.0
matplotlib>=3.7.0
gensim>=4.3.0